Time: 0 ms (100.00%), Space: 39 MB (89.63%) - LeetHub

+++ 1342-number-of-steps-to-reduce-a-number-to-zero/1342-number-of-steps-to-reduce-a-number-to-zero.java
@@ -1,7 +1,7 @@
 class Solution {
     public int numberOfSteps(int num) {
 
-        byte steps = 0;
+        short steps = 0;
         while(num != 0) {
             num = (num % 2 == 0) ? num / 2 : num - 1;
             steps++;
